[#101] Feat: 기존 Swift Data Model을 Schema Version화 시켜서 Migration 구현
