+++ CheerLot/CheerLot/Models/DTO/CheerSongItem.swift
@@ -1,0 +1,13 @@
+//
+//  CheerSongItem.swift
+//  CheerLot
+//
+//  Created by 이현주 on 9/28/25.
+//
+
+import Foundation
+
+struct CheerSongItem: Hashable {
+  let player: Player
+  let song: CheerSong
+}
