+++ CheerLot/CheerLot/Models/Domain/Schema/SchemaAliases.swift
@@ -1,0 +1,13 @@
+//
+//  SchemaAliases.swift
+//  CheerLot
+//
+//  Created by 이현주 on 9/28/25.
+//
+
+import Foundation
+
+// 현재 앱이 참조할 Schema 버전만 노출
+typealias Team = CheerLotSchemaV2.Team
+typealias Player = CheerLotSchemaV2.Player
+typealias CheerSong = CheerLotSchemaV2.CheerSong
